requirements.txt
Appium-Python-Client==2.7.1
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
et-xmlfile==1.0.1
exceptiongroup==1.0.0
h11==0.14.0
idna==2.8
jdcal==1.4.1
openpyxl==3.0.0
outcome==1.2.0
protobuf==3.15.8
pycparser==2.21
PyMySQL==0.9.3
PySocks==1.7.1
requests==2.22.0
selenium==4.5.0
six==1.15.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==1.9.4
trio==0.22.0
trio-websocket==0.9.2
typing-extensions==4.4.0
urllib3==1.26.12
wsproto==1.2.0
xlrd==1.2.0
XlsxWriter==1.2.2
xlwt==1.3.0
